+++ diffsims/tests/simulations/test_simulation.py
@@ -23,7 +23,7 @@ from diffpy.structure import Structure, Atom, Lattice
 from orix.crystal_map import Phase
 from orix.quaternion import Rotation
 
-from diffsims.simulations.simulation import Simulation
+from diffsims.simulations import Simulation2D
 from diffsims.generators.simulation_generator import SimulationGenerator
 from diffsims.crystallography.reciprocal_lattice_vector import ReciprocalLatticeVector
 
@@ -47,13 +47,13 @@ class TestSingleSimulation:
         gen = SimulationGenerator(accelerating_voltage=200)
         rot = Rotation.from_axes_angles([1, 0, 0], 45, degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0]])
-        sim = Simulation(
+        sim = Simulation2D(
             phases=al_phase, simulation_generator=gen, coordinates=coords, rotations=rot
         )
         return sim
 
     def test_init(self, single_simulation):
-        assert isinstance(single_simulation, Simulation)
+        assert isinstance(single_simulation, Simulation2D)
         assert isinstance(single_simulation.phases, Phase)
         assert isinstance(single_simulation.simulation_generator, SimulationGenerator)
         assert isinstance(single_simulation.rotations, Rotation)
@@ -97,7 +97,7 @@ class TestSimulationInitFailures:
         rot = Rotation.from_axes_angles([1, 0, 0], 45, degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0], [1, 1, 1]])
         with pytest.raises(ValueError):
-            sim = Simulation(
+            sim = Simulation2D(
                 phases=al_phase,
                 simulation_generator=gen,
                 coordinates=[coords, coords],
@@ -109,7 +109,7 @@ class TestSimulationInitFailures:
         rot = Rotation.from_axes_angles([1, 0, 0], (0, 45), degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0], [1, 1, 1]])
         with pytest.raises(ValueError):
-            sim = Simulation(
+            sim = Simulation2D(
                 phases=al_phase,
                 simulation_generator=gen,
                 coordinates=[coords, coords, coords],
@@ -121,7 +121,7 @@ class TestSimulationInitFailures:
         rot = Rotation.from_axes_angles([1, 0, 0], 45, degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0], [1, 1, 1]])
         with pytest.raises(ValueError):
-            sim = Simulation(
+            sim = Simulation2D(
                 phases=[al_phase, al_phase],
                 simulation_generator=gen,
                 coordinates=[[coords, coords], [coords, coords]],
@@ -133,7 +133,7 @@ class TestSimulationInitFailures:
         rot = Rotation.from_axes_angles([1, 0, 0], 45, degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0], [1, 1, 1]])
         with pytest.raises(ValueError):
-            sim = Simulation(
+            sim = Simulation2D(
                 phases=[al_phase, al_phase],
                 simulation_generator=gen,
                 coordinates=[[coords, coords], [coords, coords], [coords, coords]],
@@ -145,7 +145,7 @@ class TestSimulationInitFailures:
         rot = Rotation.from_axes_angles([1, 0, 0], 45, degrees=True)
         coords = ReciprocalLatticeVector(phase=al_phase, xyz=[[1, 0, 0], [1, 1, 1]])
         with pytest.raises(ValueError):
-            sim = Simulation(
+            sim = Simulation2D(
                 phases=[al_phase, al_phase],
                 simulation_generator=gen,
                 coordinates=[[coords, coords], [coords, coords], [coords, coords]],
@@ -176,7 +176,7 @@ class TestSinglePhaseMultiSimulation:
 
         vectors = [coords, coords, coords, coords]
 
-        sim = Simulation(
+        sim = Simulation2D(
             phases=al_phase,
             simulation_generator=gen,
             coordinates=vectors,
@@ -185,7 +185,7 @@ class TestSinglePhaseMultiSimulation:
         return sim
 
     def test_init(self, multi_simulation):
-        assert isinstance(multi_simulation, Simulation)
+        assert isinstance(multi_simulation, Simulation2D)
         assert isinstance(multi_simulation.phases, Phase)
         assert isinstance(multi_simulation.simulation_generator, SimulationGenerator)
         assert isinstance(multi_simulation.rotations, Rotation)
@@ -197,14 +197,14 @@ class TestSinglePhaseMultiSimulation:
 
     def test_irot(self, multi_simulation):
         sliced_sim = multi_simulation.irot[0]
-        assert isinstance(sliced_sim, Simulation)
+        assert isinstance(sliced_sim, Simulation2D)
         assert isinstance(sliced_sim.phases, Phase)
         assert sliced_sim.rotations.size == 1
         assert sliced_sim.coordinates.size == 4
 
     def test_irot_slice(self, multi_simulation):
         sliced_sim = multi_simulation.irot[0:2]
-        assert isinstance(sliced_sim, Simulation)
+        assert isinstance(sliced_sim, Simulation2D)
         assert isinstance(sliced_sim.phases, Phase)
         assert sliced_sim.rotations.size == 2
         assert sliced_sim.coordinates.size == 2
@@ -266,7 +266,7 @@ class TestMultiPhaseMultiSimulation:
         vectors = [coords, coords, coords, coords]
         al_phase2 = al_phase.deepcopy()
         al_phase2.name = "al2"
-        sim = Simulation(
+        sim = Simulation2D(
             phases=[al_phase, al_phase2],
             simulation_generator=gen,
             coordinates=[vectors, vectors],
@@ -275,7 +275,7 @@ class TestMultiPhaseMultiSimulation:
         return sim
 
     def test_init(self, multi_simulation):
-        assert isinstance(multi_simulation, Simulation)
+        assert isinstance(multi_simulation, Simulation2D)
         assert isinstance(multi_simulation.phases, np.ndarray)
         assert isinstance(multi_simulation.simulation_generator, SimulationGenerator)
         assert isinstance(multi_simulation.rotations, np.ndarray)
@@ -283,13 +283,13 @@ class TestMultiPhaseMultiSimulation:
 
     def test_iphase(self, multi_simulation):
         phase_slic = multi_simulation.iphase[0]
-        assert isinstance(phase_slic, Simulation)
+        assert isinstance(phase_slic, Simulation2D)
         assert isinstance(phase_slic.phases, Phase)
         assert phase_slic.rotations.size == 4
 
     def test_iphase_str(self, multi_simulation):
         phase_slic = multi_simulation.iphase["al"]
-        assert isinstance(phase_slic, Simulation)
+        assert isinstance(phase_slic, Simulation2D)
         assert isinstance(phase_slic.phases, Phase)
         assert phase_slic.rotations.size == 4
         assert phase_slic.phases.name == "al"
@@ -300,13 +300,13 @@ class TestMultiPhaseMultiSimulation:
 
     def test_irot(self, multi_simulation):
         sliced_sim = multi_simulation.irot[0]
-        assert isinstance(sliced_sim, Simulation)
+        assert isinstance(sliced_sim, Simulation2D)
         assert isinstance(sliced_sim.phases, np.ndarray)
         assert sliced_sim.rotations.size == 2
 
     def test_irot_slice(self, multi_simulation):
         sliced_sim = multi_simulation.irot[0:2]
-        assert isinstance(sliced_sim, Simulation)
+        assert isinstance(sliced_sim, Simulation2D)
         assert isinstance(sliced_sim.phases, np.ndarray)
         assert sliced_sim.rotations.size == 2